feat: parse daemon response
