+++ gui/src/thread.c
@@ -26,13 +26,3 @@ DOCKERCLI_CODE init_daemon(const char *path, char *args, pthread_t *tid)
 
     return DOCKERCLIE_OK;
 }
-
-DOCKERCLI_CODE check_daemon_status(void)
-{
-    int status = exec("wsl -d docker-cli curl -i -s --unix-socket /var/run/docker.sock -X GET http://localhost/containers/json > C:\\Users\\spa_1\\docker-cli\\tmp\\dockerd-status");
-    
-    if (status < 0)
-        return DOCKERCLIE_CHECKDAESTAT;
-
-    return DOCKERCLIE_OK;
-}
